fix: treat conductor properties loader as ESM
