+++ src/ductbankroute.js
@@ -1,5 +1,5 @@
 import "../workflowStatus.js";
 import "../site.js";
 import "../soilResistivityConfig.js";
-import "../conductorProperties.js";
+import "../conductorProperties.mjs";
 import "../ductbankroute.js";
